Enable basic up/down navigation within a command being edited
The way this works is that while the cursor is not on the top line,
navigating up will move the cursor up one line, while if the cursor is
on the top line, navigation will move to the previous cmd in history.

Conversely, if the cursor is not on the bottom line, navigating down will
move the cursor down one line, while if the cursor is on the bottom
line, navigation will move to the next cmd in history.